// app/types/types.ts
export type DocType =
  | "business_license"
  | "ownership_document"
  | "ecotourism_permit"
  | "national_card"
  | "bad_background_certificate";

export type InputVariantType = "outlined" | "filled" | "standard";

export type TypographyVariant =
  | 'h1'
  | 'h2'
  | 'h3'
  | 'subtitle'
  | 'body'
  | 'header'
  | 'caption';

export type BannerCategoryType = "location" | "business"
export type BannerType = "main" | "search"